shelterflow-consolidated-v2/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3==1.29.3
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pydantic==2.5.2 